images/admin-panel/requirements.txt
django==3.1
django-model-utils==4.0.0 # Набор полезных базовых классов и утилит для Django
psycopg2-binary==2.8.6
wheel==0.36.2
django_extensions==3.1.0
gunicorn==20.0.4
django-storages==1.11.1
boto3==1.17.108